dash_app/requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dash==0.28.1
dash-core-components==0.33.0
dash-html-components==0.13.2
dash-renderer==0.14.1
decorator==4.3.0
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.7
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
nbformat==4.4.0
nose==1.3.7
plotly==3.3.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
retrying==1.3.3
six==1.11.0
traitlets==4.3.2
urllib3==1.23
Werkzeug==0.14.1
